requirements.txt

django-model-utils==3.0.0
django-webpack-loader==0.5.0
jsonfield==2.0.2
requests==2.18.4
pandas==0.22.0
